% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dydx.R
\name{dydx.factor}
\alias{dydx.factor}
\title{dydx.factor}
\usage{
dydx.factor(
  data,
  model,
  variable,
  type = c("response", "link"),
  fwrap = FALSE,
  as.data.frame = TRUE,
  ...
)
}
\description{
This function is basically a copy of margins:::dydx.factor, but simulates factor level variable with a factor instead of a character. Current version of dydx.factor is problematic when predicting maxLik object with prediction.maxLik
}
