% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_factory.R
\name{gradient_factory.maxLik}
\alias{gradient_factory.maxLik}
\title{gradient_factory.maxLik}
\usage{
gradient_factory.maxLik(
  data,
  model,
  variables = NULL,
  type = "response",
  weights = NULL,
  eps = 1e-07,
  varslist = NULL,
  formula = NULL,
  ...
)
}
\description{
This function is basically a copy of margins:::gradient_factory, but introduces the formula parameter, which cannot be retrieved from maxLik object and is required to predict the data

import internal find_terms_in_model

import internal marginal_effects

import internal reset_coefs
}
