% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginal_effects.R
\name{marginal_effects.maxLik}
\alias{marginal_effects.maxLik}
\title{marginal_effects.maxLik}
\usage{
marginal_effects.maxLik(
  model,
  data = find_data(model, parent.frame()),
  variables = NULL,
  type = c("response", "link"),
  eps = 1e-07,
  varslist = NULL,
  as.data.frame = TRUE,
  ...
)
}
\description{
import internal margins find_terms_in_model

import internal margins dydx
}
