% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins.maxLik}
\alias{margins.maxLik}
\title{margins.maxLik}
\usage{
margins.maxLik(
  model,
  data = find_data(model, parent.frame()),
  variables = NULL,
  at = NULL,
  type = c("response", "link"),
  vcov = stats::vcov(model),
  vce = c("delta", "simulation", "bootstrap", "none"),
  iterations = 50L,
  unit_ses = FALSE,
  eps = 1e-07,
  formula = NULL,
  ...
)
}
\description{
import internal margins find_terms_in_model

import internal build_margins
}
