% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_maxLik.R
\name{prediction.maxLik}
\alias{prediction.maxLik}
\title{prediction.maxLik}
\usage{
\method{prediction}{maxLik}(model, data, type = "response", calculate_se = F, formula = formula)
}
\description{
prediction.maxLik
}
